fix: adapt to trace_record Address uint64 change
- Cast uint64 Address to uint32 for WASM32 memory reads
- Add go.mod replace directive for local trace_record development

+++ internal/engine/interpreter/variable_readers.go
@@ -328,7 +328,7 @@ func bytesToPointer(rawBytes []byte, typ *dwarf.PtrType, m *wasm.ModuleInstance)
 
 	// TODO: Record pointer Type info
 
-	return trace_record.ReferenceValue(dereferencedValueRecord, addr, false, typeId), typeId, nil
+	return trace_record.ReferenceValue(dereferencedValueRecord, uint64(addr), false, typeId), typeId, nil
 
 }
